refactor(rollout): centralize stock filter implementations
Keep common_filters.py as the single implementation owner while preserving
dynamic_sampling_filters.py as the compatibility path for existing configs.
