+++ miles/rollout/filter_hub/dynamic_sampling_filters.py
@@ -1,16 +1,3 @@
-import torch
-
-from miles.rollout.filter_hub.base_types import FilterOutput, iter_samples
-from miles.rollout.filter_hub.common_filters import check_no_aborted
-from miles.utils.types import Sample
+from miles.rollout.filter_hub.common_filters import check_no_aborted, check_reward_nonzero_std
 
 __all__ = ["check_reward_nonzero_std", "check_no_aborted"]
-
-
-def check_reward_nonzero_std(args, samples: list[Sample | list[Sample]], **kwargs):
-    rewards = [sample.get_reward_value(args) for sample in iter_samples(samples)]
-    keep = torch.tensor(rewards, dtype=torch.float64).std() > 1e-8
-    return FilterOutput(
-        keep=keep,
-        reason=None if keep else f"zero_std_{round(rewards[0], 1)}",
-    )
